Extract ProductsLoadedOnScroll into shared component
